backend/requirements.txt
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
av==14.4.0
bcrypt==4.3.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
coloredlogs==15.0.1
cryptography==45.0.4
ctranslate2==4.6.0
easyocr==1.7.2
faiss-cpu==1.11.0
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.5.1
greenlet==3.2.3
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
idna==3.10
imageio==2.37.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.26
langchain-core==0.3.66
langchain-text-splitters==0.3.8
langsmith==0.4.1
lazy_loader==0.4
lxml==5.4.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
onnxruntime==1.22.0
opencv-python==4.12.0.88
opencv-python-headless==4.12.0.88
orjson==3.10.18
packaging==24.2
passlib==1.7.4
pdfminer.six==20250506
pdfplumber==0.11.7
pillow==11.3.0
protobuf==6.31.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
PyMuPDF==1.26.3
PyPDF2==3.0.1
pypdfium2==4.30.1
python-bidi==0.6.6
python-docx==1.2.0
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-image==0.25.2
scipy==1.16.0
setuptools==80.9.0
shapely==2.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.6.11
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.53.0
triton==3.3.1
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
uvicorn==0.34.3
zstandard==0.23.0